refactor: remove unused files and update Dockerfile for improved structure
- Deleted obsolete files including auth_token.py, ble_pairing_helper.py, and others to streamline the codebase.
- Updated .dockerignore to exclude documentation files while retaining necessary README.md for installation.
- Modified Dockerfile to enhance installation process by using pyproject.toml and restructuring application file copying.
- Adjusted .gitignore to reflect the removal of test files and unnecessary directories.
