+++ install.ps1
@@ -550,46 +550,51 @@ function Start-Installation {
     Write-Host "INFO: Downloading application files..." -ForegroundColor Blue
     
     if ($env:LOCAL_INSTALL) {
-        # Local install for testing
+        # Local install for testing (full repo checkout)
         Write-Host "INFO: Installing from local directory: $env:LOCAL_INSTALL" -ForegroundColor Blue
-        Copy-Item "$env:LOCAL_INSTALL\packet_capture.py" $InstallDir\
-        Copy-Item "$env:LOCAL_INSTALL\auth_token.py" $InstallDir\
-        Copy-Item "$env:LOCAL_INSTALL\enums.py" $InstallDir\
-        Copy-Item "$env:LOCAL_INSTALL\ble_pairing_helper.py" $InstallDir\
+        Copy-Item "$env:LOCAL_INSTALL\pyproject.toml" $InstallDir\
+        if (Test-Path "$env:LOCAL_INSTALL\README.md") {
+            Copy-Item "$env:LOCAL_INSTALL\README.md" $InstallDir\
+        }
+        Copy-Item "$env:LOCAL_INSTALL\src" $InstallDir\src -Recurse -Force
         Copy-Item "$env:LOCAL_INSTALL\requirements.txt" $InstallDir\
         if (Test-Path "$env:LOCAL_INSTALL\.env") {
             Copy-Item "$env:LOCAL_INSTALL\.env" $InstallDir\
         }
-        # Create version info file
         New-VersionInfo
-        
         Write-Host "SUCCESS: Files copied from local directory" -ForegroundColor Green
-        }
-        else {
-        # Download from GitHub
-        Write-Host "INFO: Downloading from GitHub ($Repo @ $Branch)..." -ForegroundColor Blue
-        
-        $baseUrl = "https://raw.githubusercontent.com/$Repo/$Branch"
-        
-        # Download files
-        $files = @("packet_capture.py", "auth_token.py", "enums.py", "ble_pairing_helper.py", "ble_scan_helper.py", "requirements.txt")
-        
-        foreach ($file in $files) {
-            Write-Host "INFO: Downloading $file..." -ForegroundColor Blue
-            try {
-                Invoke-WebRequest -Uri "$baseUrl/$file" -OutFile (Join-Path $InstallDir $file) -UseBasicParsing
+    }
+    else {
+        Write-Host "INFO: Downloading repository archive from GitHub ($Repo @ $Branch)..." -ForegroundColor Blue
+        $zipUrl = "https://github.com/$Repo/archive/refs/heads/$Branch.zip"
+        $tmpZip = Join-Path $env:TEMP "meshcore-packet-capture-$Branch.zip"
+        $tmpExtract = Join-Path $env:TEMP "meshcore-packet-capture-extract-$(Get-Random)"
+        try {
+            Invoke-WebRequest -Uri $zipUrl -OutFile $tmpZip -UseBasicParsing
+            Expand-Archive -Path $tmpZip -DestinationPath $tmpExtract -Force
+            $inner = Get-ChildItem -Path $tmpExtract -Directory | Select-Object -First 1
+            if (-not $inner) {
+                throw "Archive contained no top-level directory"
             }
-            catch {
-                Write-Host "ERROR: Failed to download $file from $Repo/$Branch" -ForegroundColor Red
-                Write-Host "ERROR: Please verify the repository and branch exist" -ForegroundColor Red
-                exit 1
+            $root = $inner.FullName
+            Copy-Item "$root\pyproject.toml" $InstallDir\
+            if (Test-Path "$root\README.md") {
+                Copy-Item "$root\README.md" $InstallDir\
             }
+            Copy-Item "$root\src" $InstallDir\src -Recurse -Force
+            Copy-Item "$root\requirements.txt" $InstallDir\
         }
-        
-        # Create version info file
+        catch {
+            Write-Host "ERROR: Failed to download or extract $zipUrl" -ForegroundColor Red
+            Write-Host "ERROR: Set LOCAL_INSTALL to a full git checkout, or verify Repo/Branch." -ForegroundColor Red
+            exit 1
+        }
+        finally {
+            Remove-Item -Path $tmpZip -Force -ErrorAction SilentlyContinue
+            Remove-Item -Path $tmpExtract -Recurse -Force -ErrorAction SilentlyContinue
+        }
         New-VersionInfo
-        
-        Write-Host "SUCCESS: Files downloaded and verified" -ForegroundColor Green
+        Write-Host "SUCCESS: Repository archive installed" -ForegroundColor Green
     }
     
     # Set up virtual environment
@@ -608,7 +613,8 @@ function Start-Installation {
     & (Join-Path $InstallDir "venv\Scripts\Activate.ps1")
     & (Join-Path $InstallDir "venv\Scripts\pip.exe") install --quiet --upgrade pip
     & (Join-Path $InstallDir "venv\Scripts\pip.exe") install --quiet -r (Join-Path $InstallDir "requirements.txt")
-    Write-Host "SUCCESS: Python dependencies installed" -ForegroundColor Green
+    & (Join-Path $InstallDir "venv\Scripts\pip.exe") install --quiet $InstallDir
+    Write-Host "SUCCESS: Python dependencies and application package installed" -ForegroundColor Green
     
     # Configuration
     Write-Host ""
@@ -1331,7 +1337,7 @@ PACKETCAPTURE_MQTT1_KEEPALIVE=120
     Write-Host ""
     Write-Host "Configuration file: $InstallDir\.env.local"
     Write-Host ""
-    Write-Host "To run manually: cd $InstallDir; .\venv\Scripts\python.exe packet_capture.py"
+    Write-Host "To run manually: cd $InstallDir; .\venv\Scripts\python.exe -m meshcore_packet_capture"
     Write-Host ""
     Write-Host "SUCCESS: Installation complete!" -ForegroundColor Green
 }
