+++ pyproject.toml
@@ -1,0 +1,32 @@
+[build-system]
+requires = ["setuptools>=61", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "meshcore-packet-capture"
+description = "MeshCore Companion radio packet capture tool"
+readme = "README.md"
+requires-python = ">=3.11"
+license = { text = "MIT" }
+dynamic = ["version"]
+dependencies = [
+    "paho-mqtt>=1.6.0",
+    "meshcore>=2.2.31",
+    "bleak>=0.21.0",
+    "pyserial-asyncio>=0.6",
+    "pycayennelpp>=1.0.0",
+    "pexpect>=4.8.0",
+    "pynacl>=1.5.0",
+]
+
+[project.scripts]
+meshcore-packet-capture = "meshcore_packet_capture.__main__:cli"
+
+[tool.setuptools]
+package-dir = { "" = "src" }
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.setuptools.dynamic]
+version = { attr = "meshcore_packet_capture.__version__" }
