+++ pytest.ini
@@ -1,4 +1,3 @@
 [pytest]
-pythonpath = .
-# Hermetic unit tests (ad-hoc scripts remain under tests/ and are gitignored)
-testpaths = unit_tests
+pythonpath = src
+testpaths = tests
