+++ src/meshcore_packet_capture/__init__.py
@@ -1,0 +1,3 @@
+"""MeshCore packet capture package."""
+
+__version__ = "2.0.0"
