+++ tests/installer/test_preset_paths.py
